refactor: rename home page component, update imports
- rename home page component file to follow React PascalCase naming standards
- update import paths in index route and browser test files
- remove unused ignoreIssues entry from knip.json
