+++ knip.json
@@ -2,8 +2,5 @@
   "$schema": "https://unpkg.com/knip@6/schema.json",
   "tags": ["-lintignore"],
   "project": ["src/**/*.ts", "src/**/*.tsx", "e2e/**/*.ts", "test/**/*.ts", "test/**/*.tsx"],
-  "ignoreFiles": [],
-  "ignoreIssues": {
-    "src/data/album.ts": ["types"]
-  }
+  "ignoreFiles": []
 }
