+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-import { HomePage } from '@/components/home_page';
+import { HomePage } from '@/components/HomePage';
 
 export const Route = createFileRoute('/')({ component: HomePage });
